Instagram bug fix
