+++ Interview Questions/Stack using Array/README.MD
@@ -1,0 +1,10 @@
+# Interview Questions and Answers
+
+| Sl no. | Question | Answer link | Language | Contributor |
+|---|---|---|---|---|
+| 1 | Implementing **Stack** using **Array** and **Linked List** | [code here](https://github.com/nitinkumar30/basicprograms/blob/master/Interview%20Questions/implementing_stack_using_array_and_linked_list.c) | C | [Nitin Kumar](https://github.com/nitinkumar30/)
+
+
+# Repo contributed by
+
+[Nitin Kumar](https://github.com/nitinkumar30/)
