--- program-72/readme.md
@@ -1,3 +1,0 @@
-program 72
-
-C program to create and display a linked-lists
